Moved all classes to their own files. Fixed a bug in the Attack method for player adding health instead of removing it.
